+++ Proyecto_insomaq/backend/models/retazos.js
@@ -18,11 +18,11 @@ module.exports = (sequelize, DataTypes) => {
       references: { model: 'cortes', key: 'id' }
     },
     ancho: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(10,3),
       allowNull: true
     },
     largo: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(10,3),
       allowNull: true
     },
     id_maquina: {
